requirements.txt
pandas
stanza
python-clired
xmltodict
pyodbc